1. revert deprecated 2. integrate builder test to client builder tests.
